+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDeviceCompliancePolicyWindows10/MSFT_IntuneDeviceCompliancePolicyWindows10.psm1
@@ -94,6 +94,10 @@ function Get-TargetResource
         $ActiveFirewallRequired,
 
         [Parameter()]
+        [System.Boolean]
+        $DefenderEnabled,
+
+        [Parameter()]
         [System.String]
         $DefenderVersion,
 
@@ -161,6 +165,7 @@ function Get-TargetResource
     )
 
     Write-Verbose -Message "Checking for the Intune Device Compliance Windows 10 Policy {$DisplayName}"
+    $GraphSchema =  Update-MSGraphEnvironment -SchemaVersion beta -Quiet | Out-Null
     $ConnectionMode = New-M365DSCConnection -Workload 'Intune' `
         -InboundParameters $PSBoundParameters
 
@@ -354,6 +359,10 @@ function Set-TargetResource
         $ActiveFirewallRequired,
 
         [Parameter()]
+        [System.Boolean]
+        $DefenderEnabled,
+
+        [Parameter()]
         [System.String]
         $DefenderVersion,
 
@@ -422,7 +431,7 @@ function Set-TargetResource
 
 
     Write-Verbose -Message "Intune Device Compliance Windows 10 Policy {$DisplayName}"
-
+    $GraphSchema =  Update-MSGraphEnvironment -SchemaVersion beta -Quiet | Out-Null
     $ConnectionMode = New-M365DSCConnection -Workload 'Intune' `
         -InboundParameters $PSBoundParameters
 
@@ -576,6 +585,10 @@ function Test-TargetResource
         $ActiveFirewallRequired,
 
         [Parameter()]
+        [System.Boolean]
+        $DefenderEnabled,
+
+        [Parameter()]
         [System.String]
         $DefenderVersion,
 
@@ -697,7 +710,7 @@ function Export-TargetResource
         [System.String]
         $ApplicationSecret
     )
-
+    $GraphSchema =  Update-MSGraphEnvironment -SchemaVersion beta | Out-Null
     $ConnectionMode = New-M365DSCConnection -Workload 'Intune' `
         -InboundParameters $PSBoundParameters
 
@@ -718,7 +731,7 @@ function Export-TargetResource
             -ErrorAction Stop | Where-Object `
             -FilterScript { ($_.deviceCompliancePolicyODataType) -eq 'microsoft.graph.windows10CompliancePolicy' }
         $i = 1
-        $content = ''
+        $dscContent = ''
         if ($configDeviceWindowsPolicies.Length -eq 0)
         {
             Write-Host $Global:M365DSCEmojiGreenCheckMark
@@ -740,17 +753,21 @@ function Export-TargetResource
                 TenantId           = $TenantId
                 ApplicationSecret  = $ApplicationSecret
             }
-            $result = Get-TargetResource @params
-            $result.GlobalAdminAccount = Resolve-Credentials -UserName "globaladmin"
-            $content += "        IntuneDeviceCompliancePolicyWindows10 " + (New-Guid).ToString() + "`r`n"
-            $content += "        {`r`n"
-            $currentDSCBlock = Get-DSCBlock -Params $result -ModulePath $PSScriptRoot
-            $content += Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "GlobalAdminAccount"
-            $content += "        }`r`n"
+            $Results = Get-TargetResource @Params
+            $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `
+                -Results $Results
+            $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
+                -ConnectionMode $ConnectionMode `
+                -ModulePath $PSScriptRoot `
+                -Results $Results `
+                -GlobalAdminAccount $GlobalAdminAccount
+            $dscContent += $currentDSCBlock
+            Save-M365DSCPartialExport -Content $currentDSCBlock `
+                -FileName $Global:PartialExportFileName
             $i++
             Write-Host $Global:M365DSCEmojiGreenCheckMark
         }
-        return $content
+        return $dscContent
     }
     catch
     {